+++ algorithms/algorithms_exl.hpp
@@ -2,5 +2,6 @@
 #define ALGORITHMS_HPP
 
 #include "search.hpp"
+#include "sort.hpp"
 
 #endif // ALGORITHMS_HPP