+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Find smallest algorithm
 - Find largest algorithm
+- Selection sort algorithm
 
 ## [0.0.1] 2017-10-18